Core. Move searchResults in VisualLog.Core.Search ns
