+++ VisualLog.Core/Search/SearchEntry.cs
