+++ VisualLog.Core/Search/SearchResults.cs
@@ -1,8 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using VisualLog.Core.Search;
 
-namespace VisualLog.Core
+namespace VisualLog.Core.Search
 {
   public class SearchResults
   {
