+++ VisualLog.Core/VisualLog.Core.csproj
@@ -8,8 +8,4 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Search\" />
-  </ItemGroup>
-
 </Project>
